+++ src/dipdup/datasources/__init__.py
@@ -61,12 +61,12 @@ class Datasource(HTTPGateway, Generic[DatasourceConfigT]):
         )
         self._logger = FormattedLogger(__name__, config.name + ': {}')
 
-    @abstractmethod
-    async def run(self) -> None: ...
-
     @property
     def name(self) -> str:
         return self._config.name
+
+    async def run(self) -> None:
+        pass
 
 
 class AbiDatasource(Datasource[DatasourceConfigT], Generic[DatasourceConfigT]):
@@ -85,6 +85,9 @@ class IndexDatasource(Datasource[IndexDatasourceConfigT], Generic[IndexDatasourc
         self._on_connected_callbacks: set[EmptyCallback] = set()
         self._on_disconnected_callbacks: set[EmptyCallback] = set()
         self._on_rollback_callbacks: set[RollbackCallback] = set()
+
+    @abstractmethod
+    async def run(self) -> None: ...
 
     @abstractmethod
     async def subscribe(self) -> None: ...